+++ dlls/d2d1/d2d1_private.h
@@ -265,8 +265,6 @@ struct d2d_device_context
     struct d2d_glyph_bitmap_cache glyph_bitmap_cache;
 };
 
-BOOL d2d_device_context_prepare_reuse_target(ID2D1RenderTarget *target);
-void d2d_device_context_reset_reused_target(ID2D1RenderTarget *target, float dpi_x, float dpi_y);
 HRESULT d2d_d3d_create_render_target(struct d2d_device *device, IDXGISurface *surface, IUnknown *outer_unknown,
         const struct d2d_device_context_ops *ops, const D2D1_RENDER_TARGET_PROPERTIES *desc,
         void **render_target);
@@ -276,32 +274,16 @@ static inline BOOL d2d_device_context_is_dxgi_target(const struct d2d_device_con
     return !context->ops;
 }
 
-struct d2d_wic_cpu_glyph
-{
-    struct d2d_wic_cpu_glyph *next;
-    RECT bounds;
-    D2D1_COLOR_F colour;
-    unsigned int pitch;
-    BYTE values[];
-};
-
 struct d2d_wic_render_target
 {
     IUnknown IUnknown_iface;
     LONG refcount;
 
     IDXGISurface *dxgi_surface;
-    ID2D1Factory1 *factory;
-    ID3D10Device1 *device;
-    D2D1_RENDER_TARGET_PROPERTIES desc;
-    struct d2d_wic_render_target *pool_next;
     ID2D1RenderTarget *dxgi_target;
     IUnknown *dxgi_inner;
     ID3D10Texture2D *readback_texture;
     IWICBitmap *bitmap;
-    struct d2d_wic_cpu_glyph *cpu_glyphs;
-    struct d2d_wic_cpu_glyph **cpu_glyph_tail;
-
     unsigned int width;
     unsigned int height;
     unsigned int bpp;
@@ -309,9 +291,6 @@ struct d2d_wic_render_target
 
 HRESULT d2d_wic_render_target_init(struct d2d_wic_render_target *render_target, ID2D1Factory1 *factory,
         ID3D10Device1 *d3d_device, IWICBitmap *bitmap, const D2D1_RENDER_TARGET_PROPERTIES *desc);
-BOOL d2d_wic_render_target_reuse(ID2D1Factory1 *factory, ID3D10Device1 *device,
-        IWICBitmap *bitmap, const D2D1_RENDER_TARGET_PROPERTIES *desc, ID2D1RenderTarget **target);
-
 struct d2d_dc_render_target
 {
     ID2D1DCRenderTarget ID2D1DCRenderTarget_iface;